Move funcs to index utils.
