+++ src/index.cc
@@ -6,7 +6,6 @@
 #include <iostream>
 
 #include "minimizer_generator.h"
-#include "minimizer_utils.h"
 
 namespace chromap {
 
@@ -244,62 +243,6 @@ void Index::CheckIndex(uint32_t num_sequences,
       }
     }
   }
-}
-
-void Index::HeapMergeCandidatePositionLists(
-    const std::vector<std::vector<uint64_t>> sorted_candidate_position_lists,
-    std::vector<uint64_t> &candidate_positions) const {
-  std::priority_queue<CandidatePositionWithListIndex> heap;
-  std::vector<uint32_t> candidate_position_list_indices(
-      sorted_candidate_position_lists.size(), 0);
-
-  for (uint32_t li = 0; li < sorted_candidate_position_lists.size(); ++li) {
-    if (sorted_candidate_position_lists[li].size() == 0) {
-      continue;
-    }
-    heap.emplace(li, sorted_candidate_position_lists[li][0]);
-  }
-
-  while (!heap.empty()) {
-    const CandidatePositionWithListIndex min_candidate_position = heap.top();
-    heap.pop();
-    candidate_positions.push_back(min_candidate_position.position);
-    ++candidate_position_list_indices[min_candidate_position.list_index];
-
-    const uint32_t min_candidate_position_list_index =
-        candidate_position_list_indices[min_candidate_position.list_index];
-    const std::vector<uint64_t> &min_sorted_candidate_position_list =
-        sorted_candidate_position_lists[min_candidate_position.list_index];
-    if (min_candidate_position_list_index <
-        min_sorted_candidate_position_list.size()) {
-      heap.emplace(min_candidate_position.list_index,
-                   min_sorted_candidate_position_list
-                       [min_candidate_position_list_index]);
-    }
-  }
-}
-
-uint64_t Index::GenerateCandidatePositionForSingleSeedHit(
-    uint64_t reference_seed_hit, uint64_t read_seed_hit) const {
-  const uint32_t reference_position =
-      GenerateSequencePosition(reference_seed_hit);
-
-  const uint32_t read_position = GenerateSequencePosition(read_seed_hit);
-
-  // For now we can't see the reference here. So let us don't validate
-  // this seed hit. Instead, we do it later some time when we check the
-  // candidates.
-  const uint32_t mapping_start_position =
-      AreTwoHitsOnTheSameStrand(reference_seed_hit, read_seed_hit)
-          ? reference_position - read_position
-          : reference_position + read_position - kmer_size_ + 1;
-
-  const uint64_t reference_id = GenerateSequenceIndex(reference_seed_hit);
-
-  const uint64_t candidate_position =
-      GenerateCandidatePosition(reference_id, mapping_start_position);
-
-  return candidate_position;
 }
 
 int Index::GenerateCandidatePositions(
@@ -666,6 +609,24 @@ int Index::GenerateCandidatePositionsFromRepetitiveReadWithMateInfoOnOneStrand(
 #endif
 
   return max_minimizer_count;
+}
+
+uint64_t Index::GenerateCandidatePositionForSingleSeedHit(
+    uint64_t reference_seed_hit, uint64_t read_seed_hit) const {
+  const uint32_t reference_position =
+      GenerateSequencePosition(reference_seed_hit);
+  const uint32_t read_position = GenerateSequencePosition(read_seed_hit);
+  // For now we can't see the reference here. So let us don't validate this
+  // candidate position. Instead, we do it later some time when we check the
+  // candidates.
+  const uint32_t mapping_start_position =
+      AreTwoHitsOnTheSameStrand(reference_seed_hit, read_seed_hit)
+          ? reference_position - read_position
+          : reference_position + read_position - kmer_size_ + 1;
+  const uint64_t reference_id = GenerateSequenceIndex(reference_seed_hit);
+  const uint64_t candidate_position =
+      GenerateCandidatePosition(reference_id, mapping_start_position);
+  return candidate_position;
 }
 
 }  // namespace chromap
